+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/async/job/JobStatus.groovy
@@ -1,7 +1,7 @@
-package com.swisscom.cloud.sb.broker.backup.shield.dto
+package com.swisscom.cloud.sb.broker.async.job
 
 enum JobStatus {
-    FAILED("failed"), FINISHED("finished"), RUNNING("running")
+    FAILED("failed"), SUCCESSFUL("successful"), RUNNING("running")
 
     final String status