+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/cfapi/converter/PlanDtoConverter.groovy
@@ -1,6 +1,7 @@
 package com.swisscom.cloud.sb.broker.cfapi.converter
 
 import com.swisscom.cloud.sb.broker.cfapi.dto.PlanDto
+import com.swisscom.cloud.sb.broker.cfapi.dto.SchemasDto
 import com.swisscom.cloud.sb.broker.converter.AbstractGenericConverter
 import com.swisscom.cloud.sb.broker.model.Plan
 import groovy.transform.CompileStatic
@@ -19,6 +20,7 @@ class PlanDtoConverter extends AbstractGenericConverter<Plan, PlanDto> {
         prototype.description = source.description
         prototype.free = source.free
         prototype.metadata = convertMetadata(source)
+        prototype.schemas = new SchemasDto(source)
     }
 
     private Map<String, Object> convertMetadata(Plan plan) {
@@ -26,4 +28,5 @@ class PlanDtoConverter extends AbstractGenericConverter<Plan, PlanDto> {
         plan.metadata.each { result[it.key] = getValue(it.type, it.value) }
         result
     }
+
 }