+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/cfextensions/extensions/Status.groovy
@@ -1,0 +1,5 @@
+package com.swisscom.cloud.sb.broker.cfextensions.extensions
+
+interface Status {
+
+}