+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummyServiceProvider.groovy
@@ -3,9 +3,11 @@ package com.swisscom.cloud.sb.broker.util.test
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.google.common.base.Optional
+import com.google.gson.Gson
 import com.swisscom.cloud.sb.broker.async.AsyncProvisioningService
 import com.swisscom.cloud.sb.broker.binding.BindRequest
 import com.swisscom.cloud.sb.broker.binding.BindResponse
+import com.swisscom.cloud.sb.broker.binding.BindResponseDto
 import com.swisscom.cloud.sb.broker.binding.UnbindRequest
 import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointProvider
 import com.swisscom.cloud.sb.broker.model.*
@@ -15,25 +17,31 @@ import com.swisscom.cloud.sb.broker.provisioning.ProvisioningPersistenceService
 import com.swisscom.cloud.sb.broker.provisioning.async.AsyncOperationResult
 import com.swisscom.cloud.sb.broker.provisioning.async.AsyncServiceDeprovisioner
 import com.swisscom.cloud.sb.broker.provisioning.async.AsyncServiceProvisioner
+import com.swisscom.cloud.sb.broker.provisioning.async.AsyncServiceUpdater
 import com.swisscom.cloud.sb.broker.provisioning.job.DeprovisioningJobConfig
-import com.swisscom.cloud.sb.broker.provisioning.job.ProvisioningjobConfig
+import com.swisscom.cloud.sb.broker.provisioning.job.ProvisioningJobConfig
 import com.swisscom.cloud.sb.broker.provisioning.job.ServiceDeprovisioningJob
 import com.swisscom.cloud.sb.broker.provisioning.job.ServiceProvisioningJob
+import com.swisscom.cloud.sb.broker.provisioning.job.ServiceUpdateJob
+import com.swisscom.cloud.sb.broker.provisioning.job.UpdateJobConfig
 import com.swisscom.cloud.sb.broker.provisioning.lastoperation.LastOperationJobContext
 import com.swisscom.cloud.sb.broker.services.common.ServiceProvider
+import com.swisscom.cloud.sb.broker.updating.UpdatableProvisioner
+import com.swisscom.cloud.sb.broker.updating.UpdateResponse
 import com.swisscom.cloud.sb.broker.util.servicedetail.AbstractServiceDetailKey
 import com.swisscom.cloud.sb.broker.util.servicedetail.ServiceDetailType
 import com.swisscom.cloud.sb.broker.util.servicedetail.ServiceDetailsHelper
 import com.swisscom.cloud.sb.model.endpoint.Endpoint
 import groovy.util.logging.Slf4j
 import org.apache.commons.lang.NotImplementedException
+import org.apache.commons.lang.StringUtils
 import org.joda.time.DateTime
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
 @Component
 @Slf4j
-class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, AsyncServiceDeprovisioner, EndpointProvider {
+class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, AsyncServiceDeprovisioner, AsyncServiceUpdater, EndpointProvider, UpdatableProvisioner {
     public static final int RETRY_INTERVAL_IN_SECONDS = 10
     public static final int DEFAULT_PROCESSING_DELAY_IN_SECONDS = RETRY_INTERVAL_IN_SECONDS * 2
 
@@ -46,21 +54,39 @@ class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, 
 
     @Override
     BindResponse bind(BindRequest request) {
-        throw new NotImplementedException()
+        log.info("Bind parameters: ${request.parameters?.toString()}")
+        return new BindResponse(credentials: new BindResponseDto() {
+            @Override
+            String toJson() {
+                request.parameters ? new Gson().toJson(request.parameters) : '{}'
+            }
+        })
     }
 
     @Override
     void unbind(UnbindRequest request) {
-        throw new NotImplementedException()
+
+    }
+
+    private static DummyServiceProviderParameters DeserializeParameters(String jsonParameters) {
+        return new ObjectMapper().readValue(jsonParameters, DummyServiceProviderParameters.class)
     }
 
     @Override
     ProvisionResponse provision(ProvisionRequest request) {
+        def serviceDetails = new ArrayList<ServiceDetail>()
+        if (!StringUtils.isEmpty(request.parameters)) {
+            serviceDetails.add(new ServiceDetail(
+                    key: "mode",
+                    uniqueKey: true,
+                    value: DeserializeParameters(request.parameters).mode))
+        }
+
         if (request.acceptsIncomplete) {
-            asyncProvisioningService.scheduleProvision(new ProvisioningjobConfig(ServiceProvisioningJob.class, request, RETRY_INTERVAL_IN_SECONDS, 5))
-            return new ProvisionResponse(details: [], isAsync: true)
+            asyncProvisioningService.scheduleProvision(new ProvisioningJobConfig(ServiceProvisioningJob.class, request, RETRY_INTERVAL_IN_SECONDS, 5))
+            return new ProvisionResponse(details: serviceDetails, isAsync: true)
         } else {
-            return new ProvisionResponse(details: [], isAsync: false)
+            return new ProvisionResponse(details: serviceDetails, isAsync: false)
         }
     }
 
@@ -85,8 +111,7 @@ class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, 
             Map<String, Object> params = new ObjectMapper().readValue(context.provisionRequest.parameters, new TypeReference<Map<String, Object>>() {
             })
             if (params.containsKey('success') && !params.get('success')) {
-                return new AsyncOperationResult(status: LastOperation.Status.FAILED,
-                        details: serviceDetails)
+                return new AsyncOperationResult(status: LastOperation.Status.FAILED, details: serviceDetails)
             }
         }
 
@@ -106,7 +131,7 @@ class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, 
         return Optional.of(processOperationResultBasedOnIfEnoughTimeHasElapsed(context, delay))
     }
 
-    private boolean isServiceReady(Date dateCreation, int delayInSeconds) {
+    private static boolean isServiceReady(Date dateCreation, int delayInSeconds) {
         new DateTime(dateCreation).plusSeconds(delayInSeconds).isBeforeNow()
     }
 
@@ -115,8 +140,35 @@ class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, 
         return [new Endpoint(protocol: 'tcp', destination: '127.0.0.1', ports: '666')]
     }
 
-    enum DummyServiceProviderServiceDetailKey implements AbstractServiceDetailKey{
+    @Override
+    UpdateResponse updateParameters(UpdateRequest updateRequest) {
+        def serviceDetails = new ArrayList<ServiceDetail>()
+        if (!StringUtils.isEmpty(updateRequest.parameters)) {
+            serviceDetails.add(new ServiceDetail(
+                        key: "mode",
+                        uniqueKey: true,
+                        value: DeserializeParameters(updateRequest.parameters).mode))
+        }
 
+        if (updateRequest.acceptsIncomplete) {
+            asyncProvisioningService.scheduleUpdate(new UpdateJobConfig(ServiceUpdateJob.class, updateRequest, updateRequest.serviceInstanceGuid, RETRY_INTERVAL_IN_SECONDS, 5))
+            return new UpdateResponse(isAsync: true, details: serviceDetails)
+        } else {
+            return new UpdateResponse(isAsync: false, details: serviceDetails)
+        }
+    }
+
+    @Override
+    UpdateResponse updatePlanAndParameters(UpdateRequest updateRequest) {
+        return updateParameters(updateRequest)
+    }
+
+    @Override
+    AsyncOperationResult requestUpdate(LastOperationJobContext context) {
+        return processOperationResultBasedOnIfEnoughTimeHasElapsed(context, DEFAULT_PROCESSING_DELAY_IN_SECONDS)
+    }
+
+    enum DummyServiceProviderServiceDetailKey implements AbstractServiceDetailKey {
         DELAY_IN_SECONDS("delay_in_seconds", ServiceDetailType.OTHER)
 
         DummyServiceProviderServiceDetailKey(String key, ServiceDetailType serviceDetailType) {