+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummySynchronousBackupCapableServiceProvider.groovy
@@ -6,11 +6,14 @@ import com.swisscom.cloud.sb.broker.backup.shield.ShieldTarget
 import com.swisscom.cloud.sb.broker.binding.BindRequest
 import com.swisscom.cloud.sb.broker.binding.BindResponse
 import com.swisscom.cloud.sb.broker.binding.UnbindRequest
+import com.swisscom.cloud.sb.broker.cfextensions.extensions.Extension
 import com.swisscom.cloud.sb.broker.cfextensions.serviceusage.ServiceUsageProvider
+import com.swisscom.cloud.sb.broker.error.ErrorCode
 import com.swisscom.cloud.sb.broker.model.*
 import com.swisscom.cloud.sb.broker.provisioning.DeprovisionResponse
 import com.swisscom.cloud.sb.broker.provisioning.ProvisionResponse
 import com.swisscom.cloud.sb.broker.services.common.ServiceProvider
+import com.swisscom.cloud.sb.broker.updating.UpdateResponse
 import com.swisscom.cloud.sb.broker.util.StringGenerator
 import com.swisscom.cloud.sb.model.usage.ServiceUsage
 import com.swisscom.cloud.sb.model.usage.ServiceUsageType
@@ -79,6 +82,12 @@ class DummySynchronousBackupCapableServiceProvider implements ServiceProvider, B
 
     }
 
+    @Override
+    UpdateResponse update(UpdateRequest request) {
+        ErrorCode.SERVICE_UPDATE_NOT_ALLOWED.throwNew()
+        return null
+    }
+
     private boolean isReady(Date dateCreation) {
         new DateTime(dateCreation).plusSeconds(10).isBeforeNow()
     }
@@ -97,5 +106,10 @@ class DummySynchronousBackupCapableServiceProvider implements ServiceProvider, B
     String shieldAgentUrl(ServiceInstance serviceInstance) {
         log.info("shieldAgentUrl for ${serviceInstance.guid}")
         return StringGenerator.randomUuid()
+    }
+
+    @Override
+    Collection<Extension> buildExtensions(){
+        return [new Extension(discovery_url: "URL")]
     }
 }