+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummySynchronousServiceProvider.groovy
@@ -2,22 +2,28 @@ package com.swisscom.cloud.sb.broker.util.test
 
 import com.google.common.base.Optional
 import com.google.gson.Gson
+import com.swisscom.cloud.sb.broker.backup.shield.dto.TaskDto
 import com.swisscom.cloud.sb.broker.binding.BindRequest
 import com.swisscom.cloud.sb.broker.binding.BindResponse
 import com.swisscom.cloud.sb.broker.binding.BindResponseDto
 import com.swisscom.cloud.sb.broker.binding.UnbindRequest
+import com.swisscom.cloud.sb.broker.cfextensions.extensions.Extension
 import com.swisscom.cloud.sb.broker.cfextensions.serviceusage.ServiceUsageProvider
+import com.swisscom.cloud.sb.broker.error.ErrorCode
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.ProvisionRequest
 import com.swisscom.cloud.sb.broker.model.ServiceInstance
+import com.swisscom.cloud.sb.broker.model.UpdateRequest
 import com.swisscom.cloud.sb.broker.provisioning.DeprovisionResponse
 import com.swisscom.cloud.sb.broker.provisioning.ProvisionResponse
 import com.swisscom.cloud.sb.broker.services.common.ServiceProvider
+import com.swisscom.cloud.sb.broker.updating.UpdateResponse
 import com.swisscom.cloud.sb.model.usage.ServiceUsage
 import com.swisscom.cloud.sb.model.usage.ServiceUsageType
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import org.springframework.stereotype.Component
+import sun.reflect.generics.reflectiveObjects.NotImplementedException
 
 @Component
 @Slf4j
@@ -40,6 +46,12 @@ class DummySynchronousServiceProvider implements ServiceProvider, ServiceUsagePr
     }
 
     @Override
+    UpdateResponse update(UpdateRequest request) {
+        ErrorCode.SERVICE_UPDATE_NOT_ALLOWED.throwNew()
+        return null;
+    }
+
+    @Override
     ProvisionResponse provision(ProvisionRequest request) {
         return new ProvisionResponse(details: [], isAsync: false)
     }
@@ -53,5 +65,10 @@ class DummySynchronousServiceProvider implements ServiceProvider, ServiceUsagePr
     ServiceUsage findUsage(ServiceInstance serviceInstance, Optional<Date> enddate) {
         Date date = enddate.present ? enddate.get() : new Date()
         return new ServiceUsage(type: ServiceUsageType.TRANSACTIONS, value: "${date.time}", enddate: date)
+    }
+
+    @Override
+    Collection<Extension> buildExtensions(){
+        return [new Extension("discovery_url": "discoveryURL")]
     }
 }