+++ broker/src/main/resources/db/migration/V1_0_15__add_json_schema_columns.sql
@@ -1,0 +1,6 @@
+ALTER TABLE plan
+  ADD COLUMN service_instance_create_schema TEXT;
+ALTER TABLE plan
+  ADD COLUMN service_instance_update_schema TEXT;
+ALTER TABLE plan
+  ADD COLUMN service_binding_create_schema TEXT;