+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/cfextensions/serviceusage/ServiceUsageLookupSpec.groovy
@@ -1,19 +1,24 @@
 package com.swisscom.cloud.sb.broker.cfextensions.serviceusage
 
 import com.google.common.base.Optional
+import com.swisscom.cloud.sb.broker.backup.shield.dto.TaskDto
 import com.swisscom.cloud.sb.broker.binding.BindRequest
 import com.swisscom.cloud.sb.broker.binding.BindResponse
 import com.swisscom.cloud.sb.broker.binding.UnbindRequest
+import com.swisscom.cloud.sb.broker.cfextensions.extensions.Extension
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.ProvisionRequest
 import com.swisscom.cloud.sb.broker.model.ServiceInstance
+import com.swisscom.cloud.sb.broker.model.UpdateRequest
 import com.swisscom.cloud.sb.broker.provisioning.DeprovisionResponse
 import com.swisscom.cloud.sb.broker.provisioning.ProvisionResponse
 import com.swisscom.cloud.sb.broker.services.common.ServiceProvider
 import com.swisscom.cloud.sb.broker.services.common.ServiceProviderLookup
+import com.swisscom.cloud.sb.broker.updating.UpdateResponse
 import com.swisscom.cloud.sb.model.usage.ServiceUsage
 import com.swisscom.cloud.sb.model.usage.ServiceUsageType
 import spock.lang.Specification
+import sun.reflect.generics.reflectiveObjects.NotImplementedException
 
 class ServiceUsageLookupSpec extends Specification {
     ServiceProviderLookup serviceProviderLookup
@@ -81,6 +86,12 @@ class ServiceUsageLookupSpec extends Specification {
         }
 
         @Override
+        UpdateResponse update(UpdateRequest request) {
+            ErrorCode.SERVICE_UPDATE_NOT_ALLOWED.throwNew()
+            return null
+        }
+
+        @Override
         ProvisionResponse provision(ProvisionRequest request) {
             return null
         }
@@ -88,6 +99,14 @@ class ServiceUsageLookupSpec extends Specification {
         @Override
         DeprovisionResponse deprovision(DeprovisionRequest request) {
             return null
+        }
+
+        Collection<Extension> buildExtensions(){
+            return [new Extension("discovery_url": "discoveryURL")]
+        }
+
+        TaskDto getTask(String taskUuid){
+            new TaskDto()
         }
     }
 }