+++ client/src/main/groovy/com/swisscom/cloud/sb/client/ServiceBrokerClientExtended.groovy
@@ -84,4 +84,24 @@ class ServiceBrokerClientExtended extends ServiceBrokerClient implements IServic
         return restTemplate.exchange(appendPath("/custom/service_instances/{service_instance}/backups/{backup_id}/restores/{restore_id}"),
                 HttpMethod.GET, new HttpEntity(createSimpleAuthHeaders(cfExtUsername,cfExtPassword)), RestoreDto.class, serviceInstanceId, backupId, restoreId)
     }
+
+    @Override
+    ResponseEntity<String> getApi(String serviceInstanceId){
+        return restTemplate.exchange(appendPath("/custom/service_instances/{service_instance}/api-docs"),
+                HttpMethod.GET, new HttpEntity(createSimpleAuthHeaders(cfExtUsername,cfExtPassword)), String.class, serviceInstanceId)
+    }
+
+    //dummy endpoints
+    @Override
+    ResponseEntity<String> lockUser(String serviceInstanceId){
+        return restTemplate.exchange(appendPath("/custom/service_instances/{instanceId}/lock"),
+                HttpMethod.PUT, new HttpEntity(createSimpleAuthHeaders(cfExtUsername,cfExtPassword)), String.class, serviceInstanceId)
+    }
+
+    //dummy endpoints
+    @Override
+    ResponseEntity<String> unlockUser(String serviceInstanceId){
+        return restTemplate.exchange(appendPath("/custom/service_instances/{instanceId}/unlock"),
+                HttpMethod.PUT, new HttpEntity(createSimpleAuthHeaders(cfExtUsername,cfExtPassword)), String.class, serviceInstanceId)
+    }
 }